+++ packages/hyper_render_core/lib/src/core/render_hyper_box.dart
@@ -2,8 +2,7 @@ import 'dart:collection';
 import 'dart:math' as math;
 import 'dart:ui' as ui;
 
-import 'package:flutter/foundation.dart'
-    show debugPrint, defaultTargetPlatform, kDebugMode, TargetPlatform;
+import 'package:flutter/foundation.dart';
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';